First draft of standalone dynamic loader
